+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file.
 - getters and setters on per-face properties for Object class
 - copy method on Object class
 - set_dimensions method on Object class
+- test coverage for Object class
 
 ## [0.7.0] - 2026-03-31